requirements.txt
#DJANGO
Django==3.2.5
#DJANGO REST-FRAMEWORK
djangorestframework==3.12.4
#MYSQL
mysqlclient
PyMySQL==1.0.2
django-mysql==2.2.0
#HISTORICO DE MODELOS
django-simple-history==3.0.0
#CAPTURAR EL USER
django-crum==0.7.9
#IMAGENES
Pillow==8.3.1
#LIBRERIA DE WIDGET PARA FORMULARIOS HTMl
django-widget-tweaks==1.4.9
#FORMULARIO WIZARD
django-formtools==2.3
#crispy orm
django-crispy-forms==1.14.0
#Multiselector
django-multiselectfield==0.1.12
